Refactor based on REST collections
