+++ endpoint_user.go
@@ -1,0 +1,7 @@
+package tatsu_api
+
+import "fmt"
+
+func getUserProfile(userID string) string {
+	return baseURL + fmt.Sprintf("/users/%s/profile", userID)
+}
