+++ rest_client.go
@@ -23,119 +23,6 @@ func newRestClient(apikey string) *restClient {
 	}
 }
 
-func (rc *restClient) getAllTimeGuildMemberRanking(ctx context.Context, guildID string,
-	userID string) (*GuildMemberRanking, error) {
-	// Ensure IDs are not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	if userID == "" {
-		return nil, xerrors.New("user id was empty")
-	}
-
-	var ranking *GuildMemberRanking
-
-	// Make request.
-	err := rc.get(ctx, getGuildMemberRanking(rankingPeriodAllTime, guildID, userID), &ranking)
-
-	return ranking, err
-}
-
-func (rc *restClient) getAllTimeGuildRankings(ctx context.Context, guildID string, offset uint64) (*GuildRankings, error) {
-	// Ensure guild ID is not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	var rankings *GuildRankings
-
-	// Make request.
-	err := rc.get(ctx, getGuildRankings(rankingPeriodAllTime, guildID, offset), &rankings)
-
-	return rankings, err
-}
-
-func (rc *restClient) getCurrentMonthGuildMemberRanking(ctx context.Context, guildID string,
-	userID string) (*GuildMemberRanking, error) {
-	// Ensure IDs are not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	if userID == "" {
-		return nil, xerrors.New("user id was empty")
-	}
-
-	var ranking *GuildMemberRanking
-
-	// Make request.
-	err := rc.get(ctx, getGuildMemberRanking(rankingPeriodMonth, guildID, userID), &ranking)
-
-	return ranking, err
-}
-
-func (rc *restClient) getCurrentMonthGuildRankings(ctx context.Context, guildID string, offset uint64) (*GuildRankings, error) {
-	// Ensure guild ID is not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	var rankings *GuildRankings
-
-	// Make request.
-	err := rc.get(ctx, getGuildRankings(rankingPeriodMonth, guildID, offset), &rankings)
-
-	return rankings, err
-}
-
-func (rc *restClient) getCurrentWeekGuildMemberRanking(ctx context.Context, guildID string,
-	userID string) (*GuildMemberRanking, error) {
-	// Ensure IDs are not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	if userID == "" {
-		return nil, xerrors.New("user id was empty")
-	}
-
-	var ranking *GuildMemberRanking
-
-	// Make request.
-	err := rc.get(ctx, getGuildMemberRanking(rankingPeriodWeek, guildID, userID), &ranking)
-
-	return ranking, err
-}
-
-func (rc *restClient) getCurrentWeekGuildRankings(ctx context.Context, guildID string, offset uint64) (*GuildRankings, error) {
-	// Ensure guild ID is not empty.
-	if guildID == "" {
-		return nil, xerrors.New("guild id was empty")
-	}
-
-	var rankings *GuildRankings
-
-	// Make request.
-	err := rc.get(ctx, getGuildRankings(rankingPeriodWeek, guildID, offset), &rankings)
-
-	return rankings, err
-}
-
-func (rc *restClient) getUserProfile(ctx context.Context, userID string) (*User, error) {
-	// Ensure user ID is not empty.
-	if userID == "" {
-		return nil, xerrors.New("user id was empty")
-	}
-
-	var user *User
-
-	// Make request.
-	err := rc.get(ctx, getUserProfile(userID), &user)
-
-	return user, err
-}
-
 func (rc *restClient) get(ctx context.Context, endpoint string, v interface{}) error {
 	// Get rate limit clearance.
 	rc.bucket.acquire()
