+++ struct_error.go
@@ -1,0 +1,6 @@
+package tatsu_api
+
+type ApiError struct {
+	Code    uint   `json:"code"`
+	Message string `json:"message"`
+}
